feat(dashboard): Implement device and room management
Adds the ability to manage devices and rooms within the dashboard.
This includes fetching, adding, and displaying devices and rooms,
as well as subscribing to streams for real-time updates. Also,
database initialized and remote added new features to remote service.
This uses bloc architecture and repository pattern to handle the business
logic and data access.